Module regex_automata::dense

Types and routines specific to dense DFAs.

This module is the home of DenseDFA and each of its corresponding variant DFA types, such as Standard and ByteClass.

This module also contains a builder for configuring the construction of a dense DFA.

Structs

Builder
A builder for constructing a deterministic finite automaton from regular expressions.

ByteClass
A dense DFA that shrinks its alphabet.

Premultiplied
A dense DFA that premultiplies all of its state identifiers in its transition table.

PremultipliedByteClass
The default configuration of a dense DFA, which uses byte classes and premultiplies its state identifiers.

Standard
A standard dense DFA that does not use premultiplication or byte classes.

Enums

DenseDFA
A dense table-based deterministic finite automaton (DFA).