regex_automata::dfa

Module onepass

A DFA that can return spans for matching capturing groups.

This module is the home of a one-pass DFA.

This module also contains a Builder and a Config for building and configuring a one-pass DFA.

Structs

BuildError

An error that occurred during the construction of a one-pass DFA.

Builder

A builder for a one-pass DFA.

Cache

A cache represents mutable state that a one-pass DFA requires during a search.

Config

The configuration used for building a one-pass DFA.

DFA

A one-pass DFA for executing a subset of anchored regex searches while resolving capturing groups.

Epsilons 🔒

Epsilons represents all of the NFA epsilons transitions that went into a single transition in a single DFA state. In this case, it only represents the epsilon transitions that have some kind of non-consuming side effect: either the transition requires storing the current position of the search into a slot, or the transition is conditional and requires the current position in the input to satisfy an assertion before the transition may be taken.

InternalBuilder 🔒

An internal builder for encapsulating the state necessary to build a one-pass DFA. Typical use is just InternalBuilder::new(..).build().

PatternEpsilons 🔒

A representation of a match state’s pattern ID along with the epsilons for when a match occurs.

Slots 🔒

The set of epsilon transitions indicating that the current position in a search should be saved to a slot.

SlotsIter 🔒

An iterator over all of the bits set in a slot set.

SparseTransitionIter 🔒

An iterator over groups of consecutive equivalent transitions in a single state.

Transition 🔒

Represents a single transition in a one-pass DFA.

Enums

BuildErrorKind 🔒

The kind of error that occurred during the construction of a one-pass DFA.