regex_automata::dfa::onepass

Struct Epsilons

struct Epsilons(u64);

Epsilons represents all of the NFA epsilons transitions that went into a single transition in a single DFA state. In this case, it only represents the epsilon transitions that have some kind of non-consuming side effect: either the transition requires storing the current position of the search into a slot, or the transition is conditional and requires the current position in the input to satisfy an assertion before the transition may be taken.

This folds the cumulative effect of a group of NFA states (all connected by epsilon transitions) down into a single set of bits. While these bits can represent all possible conditional epsilon transitions, it only permits storing up to a somewhat small number of slots.

Epsilons is represented as a 42-bit integer. For example, it is packed into the lower 42 bits of a Transition. (Where the high 22 bits contains a StateID and a special โmatch winsโ property.)

Tuple Fields
0: u64
Implementations

impl Epsilons

const SLOT_MASK: u64 = 4_398_046_510_080u64

const SLOT_SHIFT: u64 = 10u64

const LOOK_MASK: u64 = 1_023u64

fn empty() -> Epsilons

Create a new empty epsilons. It has no slots and no assertions that need to be satisfied.

fn is_empty(self) -> bool

Returns true if this epsilons contains no slots and no assertions.

fn slots(self) -> Slots

Returns the slot epsilon transitions.

fn set_slots(self, slots: Slots) -> Epsilons

Set the slot epsilon transitions.

fn looks(self) -> LookSet

Return the set of look-around assertions in these epsilon transitions.

fn set_looks(self, look_set: LookSet) -> Epsilons

Set the look-around assertions on these epsilon transitions.