regex_automata::dfa::onepass

Struct PatternEpsilons

struct PatternEpsilons(u64);

A representation of a match state’s pattern ID along with the epsilons for when a match occurs.

A match state in a one-pass DFA, unlike in a more general DFA, has exactly one pattern ID. If it had more, then the original NFA would not have been one-pass.

The “epsilons” part of this corresponds to what was found in the epsilon transitions between the transition taken in the last byte of input and the ultimate match state. This might include saving slots and/or conditional epsilon transitions that must be satisfied before one can report the match.

Technically, every state has room for a ‘PatternEpsilons’, but it is only ever non-empty for match states.

Tuple Fields
0: u64
Implementations

impl PatternEpsilons

const PATTERN_ID_BITS: u64 = 22

const PATTERN_ID_SHIFT: u64

const PATTERN_ID_NONE: u64 = 0x00000000_003FFFFF

const PATTERN_ID_LIMIT: u64 = PatternEpsilons::PATTERN_ID_NONE

const PATTERN_ID_MASK: u64 = 0xFFFFFC00_00000000

const EPSILONS_MASK: u64 = 0x000003FF_FFFFFFFF

fn empty() -> PatternEpsilons

Return a new empty pattern epsilons that has no pattern ID and has no epsilons. This is suitable for non-match states.

fn is_empty(self) -> bool

Whether this pattern epsilons is empty or not. It’s empty when it has no pattern ID and an empty epsilons.

fn pattern_id(self) -> Option<PatternID>

Return the pattern ID in this pattern epsilons if one exists.

fn pattern_id_unchecked(self) -> PatternID

Returns the pattern ID without checking whether it’s valid. If this is called and there is no pattern ID in this PatternEpsilons, then this will likely produce an incorrect result or possibly even a panic or an overflow. But safety will not be violated.

This is useful when you know a particular state is a match state. If it’s a match state, then it must have a pattern ID.

fn set_pattern_id(self, pid: PatternID) -> PatternEpsilons

Return a new pattern epsilons with the given pattern ID, but the same epsilons.

fn epsilons(self) -> Epsilons

Return the epsilons part of this pattern epsilons.

fn set_epsilons(self, epsilons: Epsilons) -> PatternEpsilons

Return a new pattern epsilons with the given epsilons, but the same pattern ID.