regex_automata::meta::regex

Struct RegexInfo

pub(crate) struct RegexInfo(Arc<RegexInfoI>);

Tuple Fields
0: Arc<RegexInfoI>
Implementations

impl RegexInfo

pub(super) fn new(config: Config, hirs: &[&Hir]) -> RegexInfo

Creates a new RegexInfo from the configuration and HIRs that make up a meta regex.

This is exported for use in some tests.

pub(crate) fn config(&self) -> &Config

pub(crate) fn props(&self) -> &[Properties]

pub(crate) fn props_union(&self) -> &Properties

pub(crate) fn pattern_len(&self) -> usize

pub(crate) fn memory_usage(&self) -> usize

pub(crate) fn is_anchored_start(&self, input: &Input<'_>) -> bool

Returns true when the search is guaranteed to be anchored. That is, when a match is reported, its offset is guaranteed to correspond to the start of the search.

This includes returning true when input isn’t anchored but the underlying regex is.

pub(crate) fn is_always_anchored_start(&self) -> bool

Returns true when this regex is always anchored to the start of a search. And in particular, that regardless of an Input configuration, if any match is reported it must start at 0.

pub(crate) fn is_always_anchored_end(&self) -> bool

Returns true when this regex is always anchored to the end of a search. And in particular, that regardless of an Input configuration, if any match is reported it must end at the end of the haystack.

fn captures_disabled(&self) -> bool

Returns true when the regex’s NFA lacks capture states.

In this case, some regex engines (like the PikeVM) are unable to report match offsets, while some (like the lazy DFA can). To avoid whether a match or not is reported based on engine selection, routines that return match offsets will always report None when this is true.

Yes, this is a weird case and it’s a little fucked up. But WhichCaptures::None comes with an appropriate warning.

fn is_impossible(&self, input: &Input<'_>) -> bool

Returns true if and only if it is known that a match is impossible for the given input. This is useful for short-circuiting and avoiding running the regex engine if it’s known no match can be reported.

Note that this doesn’t necessarily detect every possible case. For example, when pattern_len() == 0, a match is impossible, but that case is so rare that it’s fine to be handled by the regex engine itself. That is, it’s not worth the cost of adding it here in order to make it a little faster. The reason is that this is called for every search. so there is some cost to adding checks here. Arguably, some of the checks that are here already probably shouldn’t be here…