Type Alias regex_automata::meta::regex::CachePool

type CachePool = Pool<Cache, Box<dyn Fn() -> Cache + Send + Sync + UnwindSafe + RefUnwindSafe>>;

A type alias for our pool of meta::Cache that fixes the type parameters to what we use for the meta regex below.

Aliased Type

struct CachePool(Box<Pool<Cache, Box<dyn Fn() -> Cache + RefUnwindSafe + Sync + Send + UnwindSafe, Global>>, Global>);

Fields
0: Box<Pool<Cache, Box<dyn Fn() -> Cache + RefUnwindSafe + Sync + Send + UnwindSafe, Global>>, Global>
Implementations

impl<T, F> Pool<T, F>

pub fn new(create: F) -> Pool<T, F>

Create a new pool. The given closure is used to create values in the pool when necessary.

impl<T: Send, F: Fn() -> T> Pool<T, F>

pub fn get(&self) -> PoolGuard<'_, T, F>

Get a value from the pool. The caller is guaranteed to have exclusive access to the given value. Namely, it is guaranteed that this will never return a value that was returned by another call to get but was not put back into the pool.

When the guard goes out of scope and its destructor is called, then it will automatically be put back into the pool. Alternatively, PoolGuard::put may be used to explicitly put it back in the pool without relying on its destructor.

Note that there is no guarantee provided about which value in the pool is returned. That is, calling get, dropping the guard (causing the value to go back into the pool) and then calling get again is not guaranteed to return the same value received in the first get call.