Type Alias regex_automata::meta::regex::CachePoolGuard

type CachePoolGuard<'a> = PoolGuard<'a, Cache, Box<dyn Fn() -> Cache + Send + Sync + UnwindSafe + RefUnwindSafe>>;

Same as above, but for the guard returned by a pool.

Aliased Type

struct CachePoolGuard<'a>(PoolGuard<'a, Cache, Box<dyn Fn() -> Cache + RefUnwindSafe + Sync + Send + UnwindSafe, Global>>);

Fields
0: PoolGuard<'a, Cache, Box<dyn Fn() -> Cache + RefUnwindSafe + Sync + Send + UnwindSafe, Global>>
Implementations

impl<'a, T: Send, F: Fn() -> T> PoolGuard<'a, T, F>

pub fn put(this: PoolGuard<'_, T, F>)

Consumes this guard and puts it back into the pool.

This circumvents the guard’s Drop implementation. This can be useful in circumstances where the automatic Drop results in poorer codegen, such as calling non-inlined functions.