regex_automata::meta

Module reverse_suffix

Routines specific to the reverse suffix optimization.

Functions

has_no_earlier_match 🔒

Returns true when it’s impossible for an earlier match to be detected after a literal candidate (corresponding to suffix) has been found.

strip_literal_suffix 🔒

Strip suffix from the end of hir and return the HIR that remains.

strip_literal_suffix_at 🔒

Strip a terminal part of suffix[..suffix_cursor] from the end of hir.

strip_repetition_literal_suffix 🔒