Module regex_automata::meta::strategy

Structs

Core ğŸ”’

Pre ğŸ”’

ReverseAnchored ğŸ”’

ReverseInner ğŸ”’

ReverseSuffix ğŸ”’

Traits

Strategy ğŸ”’

A trait that represents a single meta strategy. Its main utility is in providing a way to do dynamic dispatch over a few choices.

Functions

copy_match_to_slots ğŸ”’

Copies the offsets in the given match to the corresponding positions in slots.

new ğŸ”’