Module regex_automata::nfa::thompson::literal_trie

Structs

Frame 🔒

An explicit stack frame used for traversing the trie without using recursion.

LiteralTrie 🔒

A trie that preserves leftmost-first match semantics.

State 🔒

A state in a trie.

StateChunksIter 🔒

An iterator over all of the chunks in a state, including the active chunk.

Transition 🔒

A single transition in a trie to another state.