Module regex_automata::sparse

Types and routines specific to sparse DFAs.

This module is the home of SparseDFA and each of its corresponding variant DFA types, such as Standard and ByteClass.

Unlike the dense module, this module does not contain a builder specific for sparse DFAs. Instead, the intended way to build a sparse DFA is either by using a default configuration with its constructor, or by first configuring the construction of a dense DFA and then calling DenseDFA::to_sparse.

Structs

ByteClass
A sparse DFA that shrinks its alphabet.

Standard
A standard sparse DFA that does not use premultiplication or byte classes.

Enums

SparseDFA
A sparse table-based deterministic finite automaton (DFA).