Struct regex_automata::sparse::Standard

pub struct Standard<T: AsRef<[u8]>, S: StateID = usize>(_);

A standard sparse DFA that does not use premultiplication or byte classes.

Generally, it isn't necessary to use this type directly, since a SparseDFA can be used for searching directly. One possible reason why one might want to use this type directly is if you are implementing your own search routines by walking a DFA's transitions directly. In that case, you'll want to use this type (or any of the other DFA variant types) directly, since they implement next_state more efficiently.