Module regex_automata::util::escape

Provides convenience routines for escaping raw bytes.

Since this crate tends to deal with &[u8] everywhere and the default Debug implementation just shows decimal integers, it makes debugging those representations quite difficult. This module provides types that show &[u8] as if it were a string, with invalid UTF-8 escaped into its byte-by-byte hex representation.

Structs

DebugByte

Provides a convenient Debug implementation for a u8.

DebugHaystack

Provides a convenient Debug implementation for &[u8].