regex_automata::util::interpolate

Function find_cap_ref_braced

fn find_cap_ref_braced(rep: &[u8], i: usize) -> Option<CaptureRef<'_>>

Looks for a braced reference, e.g., ${foo1}. This assumes that an opening brace has been found at i-1 in rep. This then looks for a closing brace and returns the capture reference within the brace.