regex_automata::util::primitives

Trait IteratorIndexExt

pub(crate) trait IteratorIndexExt: Iterator {
    // Provided methods
    fn with_pattern_ids(self) -> WithPatternIDIter<Self> 
       where Self: Sized + ExactSizeIterator { ... }
    fn with_state_ids(self) -> WithStateIDIter<Self> 
       where Self: Sized + ExactSizeIterator { ... }
}

A utility trait that defines a couple of adapters for making it convenient to access indices as “small index” types. We require ExactSizeIterator so that iterator construction can do a single check to make sure the index of each element is representable by its small index type.

Provided Methods

fn with_pattern_ids(self) -> WithPatternIDIter<Self>
where Self: Sized + ExactSizeIterator,

fn with_state_ids(self) -> WithStateIDIter<Self>
where Self: Sized + ExactSizeIterator,