regex_automata::util

Module search

Types and routines that support the search APIs of most regex engines.

This sub-module isn’t exposed directly, but rather, its contents are exported at the crate root due to the universality of most of the types and routines in this module.

Structs

HalfMatch

A representation of “half” of a match reported by a DFA.

Input

The parameters for a regex search including the haystack to search.

Match

A representation of a match reported by a regex engine.

MatchError

An error indicating that a search stopped before reporting whether a match exists or not.

PatternSet

A set of PatternIDs.

PatternSetInsertError

An error that occurs when a PatternID failed to insert into a PatternSet.

PatternSetIter

An iterator over all pattern identifiers in a PatternSet.

Span

A representation of a span reported by a regex engine.

Enums

Anchored

The type of anchored search to perform.

MatchErrorKind

The underlying kind of a MatchError.

MatchKind

The kind of match semantics to use for a regex pattern.