Function regex_automata::util::syntax::parse

pub fn parse(pattern: &str) -> Result<Hir, Error>

A convenience routine for parsing a pattern into an HIR value with the default configuration.

Example

This shows how to parse a pattern into an HIR value:

use regex_automata::util::syntax;

let hir = syntax::parse(r"([a-z]+)|([0-9]+)")?;
assert_eq!(Some(1), hir.properties().static_explicit_captures_len());