Crate regex_automata

This crate provides an “expert” API for executing regular expressions using finite automata.

WARNING: This 0.2 release of regex-automata was published before it was ready to unblock work elsewhere that needed some of the new APIs in this release. At the time of writing, it is strongly preferred that you continue using the regex-automata 0.1 release. Since this release represents an unfinished state, please do not create issues for this release unless it’s for a critical bug.

Modules

dfa

A module for building and searching with determinstic finite automata (DFAs).

Structs

DeserializeError

An error that occurs when deserializing an object defined in this crate.

HalfMatch

A representation of a match reported by a DFA.

Match

A representation of a match reported by a regex engine.

MultiMatch

A representation of a multi match reported by a regex engine.

PatternID

An identifier for a regex pattern.

SerializeError

An error that occurs when serializing an object from this crate.

Enums

MatchError

An error type indicating that a search stopped prematurely without finding a match.

MatchKind

The kind of match semantics to use for a DFA.