Module regex_automata::dfa::onepass

A DFA that can return spans for matching capturing groups.

This module is the home of a one-pass DFA.

This module also contains a Builder and a Config for building and configuring a one-pass DFA.

Structs

BuildError

An error that occurred during the construction of a one-pass DFA.

Builder

A builder for a one-pass DFA.

Cache

A cache represents mutable state that a one-pass DFA requires during a search.

Config

The configuration used for building a one-pass DFA.

DFA

A one-pass DFA for executing a subset of anchored regex searches while resolving capturing groups.