Module regex_automata::hybrid::regex

Available on crate feature hybrid only.

A lazy DFA backed Regex.

This module provides a Regex backed by a lazy DFA. A Regex implements convenience routines you might have come to expect, such as finding a match and iterating over all non-overlapping matches. This Regex type is limited in its capabilities to what a lazy DFA can provide. Therefore, APIs involving capturing groups, for example, are not provided.

Internally, a Regex is composed of two DFAs. One is a “forward” DFA that finds the end offset of a match, where as the other is a “reverse” DFA that find the start offset of a match.

See the parent module for examples.

Structs

Builder

A builder for a regex based on a hybrid NFA/DFA.

Cache

A cache represents a partially computed forward and reverse DFA.

FindMatches

An iterator over all non-overlapping matches for an infallible search.

Regex

A regular expression that uses hybrid NFA/DFAs (also called “lazy DFAs”) for searching.