Module regex_automata::meta::regex

Structs

Builder

A builder for configuring and constructing a Regex.

Cache

Represents mutable scratch space used by regex engines during a search.

CapturesMatches

An iterator over all non-overlapping leftmost matches with their capturing groups.

Config

An object describing the configuration of a Regex.

FindMatches

An iterator over all non-overlapping matches.

Regex

A regex matcher that works by composing several other regex matchers automatically.

RegexI 🔒

The internal implementation of Regex, split out so that it can be wrapped in an Arc.

RegexInfo 🔒

RegexInfoI 🔒

Split

Yields all substrings delimited by a regular expression match.

SplitN

Yields at most N spans delimited by a regular expression match.

Type Aliases

CachePool 🔒

A type alias for our pool of meta::Cache that fixes the type parameters to what we use for the meta regex below.

CachePoolFn 🔒

The type of the closure we use to create new caches. We need to spell out all of the marker traits or else we risk leaking !MARKER impls.

CachePoolGuard 🔒

Same as above, but for the guard returned by a pool.