Module regex_automata::nfa::thompson::compiler

Structs

Compiler

A builder for compiling an NFA from a regex’s high-level intermediate representation (HIR).

Config

The configuration used for a Thompson NFA compiler.

ThompsonRef 🔒

A value that represents the result of compiling a sub-expression of a regex’s HIR. Specifically, this represents a sub-graph of the NFA that has an initial state at start and a final state at end.

Utf8Compiler 🔒

A UTF-8 compiler based on Daciuk’s algorithm for compilining minimal DFAs from a lexicographically sorted sequence of strings in linear time.

Utf8LastTransition 🔒

Utf8Node 🔒

Utf8State 🔒

Enums

WhichCaptures

A configuration indicating which kinds of State::Capture states to include.