regex_automata::nfa::thompson

Module pikevm

Available on crate features nfa-thompson and nfa-pikevm only.

An NFA backed Pike VM for executing regex searches with capturing groups.

This module provides a PikeVM that works by simulating an NFA and resolving all spans of capturing groups that participate in a match.

Structs

Builder

A builder for a PikeVM.

Cache

A cache represents mutable state that a PikeVM requires during a search.

CapturesMatches

An iterator over all non-overlapping leftmost matches, with their capturing groups, for a particular search.

Config

The configuration used for building a PikeVM.

FindMatches

An iterator over all non-overlapping matches for a particular search.

PikeVM

A virtual machine for executing regex searches with capturing groups.