Module regex_automata::util::iter

Generic helpers for iteration of matches from a regex engine in a haystack.

The principle type in this module is a Searcher. A Searcher provides its own lower level iterator-like API in addition to methods for constructing types that implement Iterator. The documentation for Searcher explains a bit more about why these different APIs exist.

Currently, this module supports iteration over any regex engine that works with the HalfMatch, Match or Captures types.

Structs

CapturesIter

An iterator over all non-overlapping captures for an infallible search.

HalfMatchesIter

An iterator over all non-overlapping half matches for an infallible search.

MatchesIter

An iterator over all non-overlapping matches for an infallible search.

Searcher

A searcher for creating iterators and performing lower level iteration.

TryCapturesIter

An iterator over all non-overlapping captures for a fallible search.

TryHalfMatchesIter

An iterator over all non-overlapping half matches for a fallible search.

TryMatchesIter

An iterator over all non-overlapping matches for a fallible search.