regex_automata::util

Module look

Types and routines for working with look-around assertions.

This module principally defines two types:

Look enumerates all of the assertions supported by this crate.

LookSet provides a way to efficiently store a set of Look values.

LookMatcher provides routines for checking whether a Look or a LookSet matches at a particular position in a haystack.

Structs

LookMatcher

A matcher for look-around assertions.

LookSet

LookSet is a memory-efficient set of look-around assertions.

LookSetIter

An iterator over all look-around assertions in a LookSet.

UnicodeWordBoundaryError

An error that occurs when the Unicode-aware \w class is unavailable.

Enums

Look

A look-around assertion.