regex_automata::util

Module syntax

Available on crate feature syntax only.

Utilities for dealing with the syntax of a regular expression.

This module currently only exposes a Config type that itself represents a wrapper around the configuration for a regex-syntax::ParserBuilder. The purpose of this wrapper is to make configuring syntax options very similar to how other configuration is done throughout this crate. Namely, instead of duplicating syntax options across every builder (of which there are many), we instead create small config objects like this one that can be passed around and composed.

Structs

Config

A common set of configuration options that apply to the syntax of a regex.

Functions

parse

A convenience routine for parsing a pattern into an HIR value with the default configuration.

parse_many

A convenience routine for parsing many patterns into HIR value with the default configuration.

parse_many_with

A convenience routine for parsing many patterns into HIR values using a Config.

parse_with

A convenience routine for parsing a pattern into an HIR value using a Config.