Function regex_automata::util::syntax::parse_with

pub fn parse_with(pattern: &str, config: &Config) -> Result<Hir, Error>

Available on crate feature syntax only.

A convenience routine for parsing a pattern into an HIR value using a Config.

Example

This shows how to parse a pattern into an HIR value with a non-default configuration:

use regex_automata::util::syntax;

let hir = syntax::parse_with(
    r"^[a-z]+$",
    &syntax::Config::new().multi_line(true).crlf(true),
)?;
assert!(hir.properties().look_set().contains_anchor_crlf());