Module regex_automata::util::wire

Types and routines that support the wire format of finite automata.

Currently, this module just exports a few error types and some small helpers for deserializing dense DFAs using correct alignment.

Structs

AlignAs

A hack to align a smaller type B with a bigger type T.

DeserializeError

An error that occurs when deserializing an object defined in this crate.

SerializeError

An error that occurs when serializing an object from this crate.